Move and update Identity transform
