+++ monai/transforms/spatial/array.py
@@ -36,18 +36,6 @@ from monai.transforms.utils import (
 from monai.utils.misc import ensure_tuple
 
 
-class Identity(Transform):
-    """
-    Identity transform output is same as the input.
-    """
-
-    def __init__(self):
-        pass
-
-    def __call__(self, data):
-        return data
-
-
 class Spacing(Transform):
     """
     Resample input image into the specified `pixdim`.
